Repackaging the library
