--- scoreboard-core/src/main/java/io/github/filipchrzescijanek/MatchesContainer.java
@@ -1,7 +1,0 @@
-package io.github.filipchrzescijanek;
-
-public interface MatchesContainer {
-
-    Match getById(String id);
-
-}
