+++ scoreboard-core/src/main/java/io/github/filipchrzescijanek/scoreboard/queries/GetMatchByIdHandler.java
@@ -1,0 +1,9 @@
+package io.github.filipchrzescijanek.scoreboard.queries;
+
+import io.github.filipchrzescijanek.scoreboard.domain.Match;
+
+public interface GetMatchByIdHandler {
+
+    Match getById(String id);
+
+}
