+++ scoreboard-core/src/test/java/io/github/filipchrzescijanek/scoreboard/commands/AddMatchHandlerTest.java
@@ -1,9 +1,13 @@
-package io.github.filipchrzescijanek;
+package io.github.filipchrzescijanek.scoreboard.commands;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import io.github.filipchrzescijanek.scoreboard.core.Scoreboard;
+import io.github.filipchrzescijanek.scoreboard.domain.Match;
+import io.github.filipchrzescijanek.scoreboard.domain.Score;
 
 import static org.assertj.core.api.Assertions.*;
